[RENAME] auth.type.ts
type 단수형으로 표시
