+++ src/shared/authentication/auth.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-import { AuthGuardRequest, TokenPayload } from "./auth.types";
+import { AuthGuardRequest, TokenPayload } from "./auth.type";
 
 export const Auth = createParamDecorator((data: keyof TokenPayload, context: ExecutionContext) => {
   const req: AuthGuardRequest = context.switchToHttp().getRequest();
