+++ src/shared/authentication/auth.type.ts
